+++ react/CHANGELOG.md
@@ -17,6 +17,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [8.0.1] - 2020-06-29
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+* Make `ansi-to-html` a runtime dependency
+
 ## [8.0.0] - 2020-06-29
 
 ### Added
@@ -24,6 +38,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Display execution summary ([#1067](https://github.com/cucumber/cucumber/pull/1067))
 * Display failed Hooks and attachments added in Hooks ([#975](https://github.com/cucumber/cucumber/pull/975))
 * Use ANSI color to render logs ([#1057](https://github.com/cucumber/cucumber/issues/1057))
+* Add search ([#895](https://github.com/cucumber/cucumber/pull/895))
 
 ### Changed
 
@@ -179,7 +194,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v8.0.0...master
+[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v8.0.1...master
+[8.0.1]:      https://github.com/cucumber/cucumber/compare/react/v8.0.0...react/v8.0.1
 [8.0.0]:      https://github.com/cucumber/cucumber/compare/react/v7.0.0...react/v8.0.0
 [7.0.0]:      https://github.com/cucumber/cucumber/compare/react/v6.0.0...react/v7.0.0
 [6.0.0]:      https://github.com/cucumber/cucumber/compare/react/v5.1.0...react/v6.0.0